Refactor Gistbook output view.
Managers have been renamed to Services. Services have been placed in the output-view
directory. The views in the separate output directory were moved into the output-view
directory.

Resolves #209
