+++ client.src/shared/views/gistbook-view/views/output-view/services/module-bundler.js
