+++ client.src/shared/views/gistbook-view/views/output-view/views/document-view/document-view.jst
